% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataItemExplorer.R
\name{dataItemExplorer}
\alias{dataItemExplorer}
\title{Data Item Explorer API}
\usage{
dataItemExplorer(dataitems, fromdate = Sys.Date() - 30, todate = Sys.Date(),
  datetype = "gasday", latestflag = "Y", applicableforflag = "Y",
  apiurl = "http://marketinformation.natgrid.co.uk/MIPIws-public/public/publicwebservice.asmx",
  http_proxy = NULL)
}
\arguments{
\item{dataitems}{A vector of characters containing data items to enquire via API.}

\item{fromdate}{A character object specifying the start date. Date is inclusive.}

\item{todate}{A character object specifying the end date. Date is inclusive.}

\item{datetype}{A character object specifying the data type. Defaults to \code{gasday}}

\item{latestflag}{A character object with length of one to specify whether to extract the latest data. This can either be \code{Y} or \code{N}. Defaults to \code{Y}.}

\item{applicableforflag}{A character object with length of one to specify whether dates specified are 'applicable for' or 'applicable on'. This can either be \code{Y} or \code{N} where \code{Y} indicates 'applicable for'. Defaults to \code{Y}.}

\item{apiurl}{A character object which points to National Grid's SOAP API. Defaults to 'http://marketinformation.natgrid.co.uk/MIPIws-public/public/publicwebservice.asmx'}

\item{http_proxy}{A character object which points to your proxy server. This will assign the specified value to environment variable \code{http_proxy}. It will not assign anything if the value is \code{NULL}. Defaults to \code{NULL}.}
}
\value{
A dataframe object
}
\description{
This function connects to the UK National Grid's API for Data Item Explorer, which is a major data source for gas-related information. Internet connection must be available.
}
\details{
The function submits an enquiry to the destination API using XML over HTTP using SOAP standard. The HTTP response is in XML format which function will parse internally and return a R dataframe object.
}
\examples{
require(ggplot2)

# Specify the data item(s) to enquire from API
dataitems <- c('Storage Injection, Actual','Storage Withdrawal, Actual')

# Initialise API (requires internet connection for this step)
response <- dataItemExplorer(dataitems,fromdate = '2013-10-01')

# Converts the results to appropiate types
response$ApplicableFor <- as.Date(response$ApplicableFor)
response$Value <- as.numeric(response$Value)

# Visualise the results on a chart
ggplot(response,aes(x=ApplicableFor,y=Value,colour=PublicationObjectName)) + geom_line()
}
\author{
Timothy Wong, \email{timothy.wong@hotmail.co.uk}
}
\references{
\itemize{
\item Graphical User Interface for Data Item Explorer\cr
\url{http://www2.nationalgrid.com/uk/industry-information/gas-transmission-operational-data/data-item-explorer/}
\item API specification\cr
\url{http://marketinformation.natgrid.co.uk/MIPIws-public/public/publicwebservice.asmx?op=GetPublicationDataWM}
}
}

